requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
boto3 python-multipart